imka-ridge-regression/requirements.txt
aihwkit==0.9.0
dill==0.3.8
easydict==1.10
hadamard_transform==0.1.3
matplotlib==3.6.3
numpy==1.23.5
PyYAML==6.0.1
scikit_learn==1.0.2
scipy==1.12.0
setuptools==65.6.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
